Change in host common utility and corresponding updates required for Reduction
