+++ test/reduce/reduce_no_index.cpp
@@ -6,7 +6,7 @@
 #include "host_tensor.hpp"
 #include "host_tensor_generator.hpp"
 #include "host_reduction.hpp"
-#include "reduce_util.hpp"
+#include "host_common_util.hpp"
 
 using namespace ck;
 
@@ -72,6 +72,7 @@ bool test_reduce_no_index_impl(int init_method,
     using namespace ck::tensor_operation::device;
     using namespace ck::tensor_operation::device::device_reduce_instance;
     using namespace ck::host_reduce;
+    using ck::host_common::to_int_vector;
 
     constexpr bool out_support_atomic_add = std::is_same<OutDataType, float>::value;
     constexpr bool op_support_atomic_add  = true;
@@ -91,10 +92,6 @@ bool test_reduce_no_index_impl(int init_method,
 
     Tensor<OutDataType> out_ref(outLengths);
     Tensor<OutDataType> out(outLengths);
-
-    // only used when the OutDataType is bhalf_t
-    Tensor<float> out_ref_fp32(outLengths);
-    Tensor<float> out_fp32(outLengths);
 
     auto inStrides  = in.mDesc.GetStrides();
     auto outStrides = out.mDesc.GetStrides();
@@ -285,19 +282,8 @@ bool test_reduce_no_index_impl(int init_method,
 
         bool single_result = true;
 
-        if constexpr(std::is_same<OutDataType, ck::half_t>::value ||
-                     std::is_same<OutDataType, ck::bhalf_t>::value)
-        {
-            reduce_util::to_f32_vector(out, out_fp32);
-            reduce_util::to_f32_vector(out_ref, out_ref_fp32);
-            single_result = ck::utils::check_err(
-                out_fp32.mData, out_ref_fp32.mData, "Error: incorrect data result!");
-        }
-        else
-        {
-            single_result =
-                ck::utils::check_err(out.mData, out_ref.mData, "Error: incorrect data result!");
-        };
+        single_result =
+            ck::utils::check_err(out.mData, out_ref.mData, "Error: incorrect data result!");
 
         if(!single_result)
         {
@@ -372,19 +358,8 @@ bool test_reduce_no_index_impl(int init_method,
 
             bool single_result = true;
 
-            if constexpr(std::is_same<OutDataType, ck::half_t>::value ||
-                         std::is_same<OutDataType, ck::bhalf_t>::value)
-            {
-                reduce_util::to_f32_vector(out, out_fp32);
-                reduce_util::to_f32_vector(out_ref, out_ref_fp32);
-                single_result = ck::utils::check_err(
-                    out_fp32.mData, out_ref_fp32.mData, "Error: incorrect data result!");
-            }
-            else
-            {
-                single_result =
-                    ck::utils::check_err(out.mData, out_ref.mData, "Error: incorrect data result!");
-            };
+            single_result =
+                ck::utils::check_err(out.mData, out_ref.mData, "Error: incorrect data result!");
 
             if(!single_result)
             {
